% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bounding_box_polygon}
\alias{bounding_box_polygon}
\title{Bounding Box Polygon}
\format{
A simple feature collection with 1 feature and 1 field.
}
\usage{
bounding_box_polygon
}
\description{
A sample sf object representing the study area boundary.
}
\keyword{datasets}
