% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_polygons}
\alias{input_polygons}
\title{Input Polygons}
\format{
A simple feature collection with multiple features.
}
\usage{
input_polygons
}
\description{
A sample sf object representing the first set of polygons (e.g., artifacts, sites).
}
\keyword{datasets}
