% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_polygons}
\alias{other_polygons}
\title{Other Polygons}
\format{
A simple feature collection with multiple features.
}
\usage{
other_polygons
}
\description{
A sample sf object representing the second set of polygons for overlap comparison.
}
\keyword{datasets}
