% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_overlap_distribution}
\alias{plot_overlap_distribution}
\title{Plot overlap distribution with observed value}
\usage{
plot_overlap_distribution(areas_df, observed_value, n, pval)
}
\arguments{
\item{areas_df}{data.frame with 'area' column from random shuffles}

\item{observed_value}{numeric observed overlap area}

\item{n}{number of shuffles}

\item{pval}{numeric p-value}
}
\description{
Plot overlap distribution with observed value
}
