% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_polygons}
\alias{plot_polygons}
\title{Plot polygons for visualization}
\usage{
plot_polygons(
  bounding_box_polygon,
  input_polygons,
  other_polygons,
  title = "Plot of input polygons"
)
}
\arguments{
\item{bounding_box_polygon}{sf object}

\item{input_polygons}{sf object}

\item{other_polygons}{sf object}

\item{title}{character string for plot title}
}
\description{
Plot polygons for visualization
}
