% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{polygonoverlap-package}
\alias{polygonoverlap}
\alias{polygonoverlap-package}
\title{Polygon Overlap Analysis using sf}
\description{
Modern implementation using sf package instead of sp/rgeos/rgdal
}
\details{
This package computes the probability that an observed area of overlap
between two sets of polygons is due to chance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benmarwick/polygonoverlap}
  \item Report bugs at \url{https://github.com/benmarwick/polygonoverlap/issues}
}

}
\keyword{internal}
