% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{safe_read_shapefile}
\alias{safe_read_shapefile}
\title{Helper function to safely read shapefiles}
\usage{
safe_read_shapefile(path, verbose = FALSE)
}
\arguments{
\item{path}{path to shapefile}

\item{verbose}{whether to print messages}
}
\description{
Helper function to safely read shapefiles
}
