% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compute_overlap_area_of_polygons_observed}
\alias{compute_overlap_area_of_polygons_observed}
\title{Compute overlap area of polygons observed}
\usage{
compute_overlap_area_of_polygons_observed(input_polygons, other_polygons)
}
\arguments{
\item{input_polygons}{SpatialPolygonsDataFrame, many polygons}

\item{other_polygons}{SpatialPolygonsDataFrame, many polygons}
}
\description{
Compute overlap area of polygons observed
}
