% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{shift_poly_to_random_points}
\alias{shift_poly_to_random_points}
\title{Shift polygons to random points}
\usage{
shift_poly_to_random_points(bounding_box_polygon, input_polygons,
  n = 100)
}
\arguments{
\item{bounding_box_polygon}{SpatialPolygonsDataFrame, a single polygon bounding box}

\item{input_polygons}{SpatialPolygonsDataFrame, many polygons}

\item{n}{integer, 100 to 1000 are usually good values}
}
\description{
Shift polygons to random points in a bounding box
}
